+++ src/pixel.h
@@ -6,9 +6,9 @@
 #include <string>
 
 struct Color {
-  double r;
-  double g;
-  double b;
+  float r;
+  float g;
+  float b;
 
   Color operator+(const Color& other) const;
   Color operator-(const Color& other) const;
@@ -17,8 +17,8 @@ struct Color {
   bool operator==(const Color& other) const;
   bool operator!=(const Color& other) const;
 
-  Color operator*(const double scalar) const;
-  Color operator/(const double scalar) const;
+  Color operator*(const float scalar) const;
+  Color operator/(const float scalar) const;
 
   bool IsColorInRange() const;
   const char* ToHex() const;
